refactor: Sidebar 컴포넌트 분리
- 여정 정보 입력 섹션
- 내가 갈 곳 섹션
